Revert "demo email campaign"
