+++ utils/testflow/gentool/go.mod
@@ -6,7 +6,7 @@ require (
 	github.com/gobuffalo/flect v1.0.2
 	github.com/pkg/errors v0.9.1
 	github.com/qor5/admin/v3 v3.0.1-0.20240424102851-d75759576158
-	github.com/qor5/web/v3 v3.0.12-0.20250108021443-325c9ca76ac5
+	github.com/qor5/web/v3 v3.0.12-0.20241218104310-f11d78471a09
 	github.com/sergi/go-diff v1.3.1
 	mvdan.cc/gofumpt v0.6.0
 )
